+++ src/main/java/uk/gov/hmcts/divorce/divorcecase/model/CaseData.java
@@ -20,22 +20,14 @@ import uk.gov.hmcts.divorce.divorcecase.model.access.CaseworkerAccess;
 import uk.gov.hmcts.divorce.divorcecase.model.access.CaseworkerAccessBetaOnlyAccess;
 import uk.gov.hmcts.divorce.divorcecase.model.access.CaseworkerAndSuperUserAccess;
 import uk.gov.hmcts.divorce.divorcecase.model.access.CaseworkerBulkScanAccess;
-import uk.gov.hmcts.divorce.divorcecase.model.access.CaseworkerCourtAdminWithSolicitorAccess;
 import uk.gov.hmcts.divorce.divorcecase.model.access.DefaultAccess;
-import uk.gov.hmcts.divorce.document.model.ConfidentialDivorceDocument;
 import uk.gov.hmcts.divorce.document.model.DivorceDocument;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.UUID;
 
 import static java.util.Objects.nonNull;
-import static java.util.stream.Collectors.groupingBy;
-import static java.util.stream.Collectors.toCollection;
 import static org.springframework.util.CollectionUtils.isEmpty;
 import static uk.gov.hmcts.ccd.sdk.type.FieldType.CasePaymentHistoryViewer;
 import static uk.gov.hmcts.ccd.sdk.type.FieldType.Collection;
@@ -157,14 +149,6 @@ public class CaseData {
         access = {CaseworkerAccessBetaOnlyAccess.class}
     )
     private DivorceDocument d11Document;
-
-    @CCD(
-        label = "Confidential documents uploaded",
-        typeOverride = Collection,
-        typeParameterOverride = "ConfidentialDivorceDocument",
-        access = {CaseworkerCourtAdminWithSolicitorAccess.class}
-    )
-    private List<ListValue<ConfidentialDivorceDocument>> confidentialDocumentsUploaded;
 
     @CCD(
         label = "General Orders",
@@ -306,52 +290,6 @@ public class CaseData {
         }
 
         return applicant2Email;
-    }
-
-    public void sortConfidentialDocuments(List<ListValue<ConfidentialDivorceDocument>> previousDocuments) {
-        if (isEmpty(previousDocuments)) {
-            return;
-        }
-
-        Set<String> previousListValueIds = previousDocuments
-            .stream()
-            .map(ListValue::getId)
-            .collect(toCollection(HashSet::new));
-
-        //Split the collection into two lists one without id's(newly added documents) and other with id's(existing documents)
-        Map<Boolean, List<ListValue<ConfidentialDivorceDocument>>> documentsWithoutIds =
-            this.getConfidentialDocumentsUploaded()
-                .stream()
-                .collect(groupingBy(listValue -> !previousListValueIds.contains(listValue.getId())));
-
-        List<ListValue<ConfidentialDivorceDocument>> sortedDocuments = new ArrayList<>();
-        sortedDocuments.addAll(0, documentsWithoutIds.get(true)); // add new documents to start of the list
-        sortedDocuments.addAll(1, documentsWithoutIds.get(false));
-
-        sortedDocuments.forEach(
-            uploadedDocumentListValue -> uploadedDocumentListValue.setId(String.valueOf(UUID.randomUUID()))
-        );
-
-        this.setConfidentialDocumentsUploaded(sortedDocuments);
-    }
-
-    private List<ListValue<DivorceDocument>> sortDocuments(final Map<Boolean, List<ListValue<DivorceDocument>>> documentsWithoutIds) {
-
-        final List<ListValue<DivorceDocument>> sortedDocuments = new ArrayList<>();
-
-        final var newDocuments = documentsWithoutIds.get(true);
-        final var previousDocuments = documentsWithoutIds.get(false);
-
-        if (null != newDocuments) {
-            sortedDocuments.addAll(0, newDocuments); // add new documents to start of the list
-            sortedDocuments.addAll(1, previousDocuments);
-            sortedDocuments.forEach(
-                uploadedDocumentListValue -> uploadedDocumentListValue.setId(String.valueOf(UUID.randomUUID()))
-            );
-            return sortedDocuments;
-        }
-
-        return previousDocuments;
     }
 
     public void archiveAlternativeServiceApplicationOnCompletion() {
