+++ src/main/java/uk/gov/hmcts/divorce/divorcecase/model/CaseDocuments.java
@@ -7,7 +7,9 @@ import lombok.NoArgsConstructor;
 import uk.gov.hmcts.ccd.sdk.api.CCD;
 import uk.gov.hmcts.ccd.sdk.type.ListValue;
 import uk.gov.hmcts.divorce.divorcecase.model.access.Applicant2Access;
+import uk.gov.hmcts.divorce.divorcecase.model.access.CaseworkerCourtAdminWithSolicitorAccess;
 import uk.gov.hmcts.divorce.divorcecase.model.access.DefaultAccess;
+import uk.gov.hmcts.divorce.document.model.ConfidentialDivorceDocument;
 import uk.gov.hmcts.divorce.document.model.DivorceDocument;
 
 import java.util.ArrayList;
@@ -60,6 +62,14 @@ public class CaseDocuments {
     )
     private List<ListValue<DivorceDocument>> documentsUploaded;
 
+    @CCD(
+        label = "Confidential documents uploaded",
+        typeOverride = Collection,
+        typeParameterOverride = "ConfidentialDivorceDocument",
+        access = {CaseworkerCourtAdminWithSolicitorAccess.class}
+    )
+    private List<ListValue<ConfidentialDivorceDocument>> confidentialDocumentsUploaded;
+
     public void sortApplicant1UploadedDocuments(List<ListValue<DivorceDocument>> previousDocuments) {
         if (isEmpty(previousDocuments)) {
             return;
@@ -96,6 +106,33 @@ public class CaseDocuments {
                 .collect(groupingBy(listValue -> !previousListValueIds.contains(listValue.getId())));
 
         this.setDocumentsUploaded(sortDocuments(documentsWithoutIds));
+    }
+
+    public void sortConfidentialDocuments(List<ListValue<ConfidentialDivorceDocument>> previousDocuments) {
+        if (isEmpty(previousDocuments)) {
+            return;
+        }
+
+        Set<String> previousListValueIds = previousDocuments
+            .stream()
+            .map(ListValue::getId)
+            .collect(toCollection(HashSet::new));
+
+        //Split the collection into two lists one without id's(newly added documents) and other with id's(existing documents)
+        Map<Boolean, List<ListValue<ConfidentialDivorceDocument>>> documentsWithoutIds =
+            this.getConfidentialDocumentsUploaded()
+                .stream()
+                .collect(groupingBy(listValue -> !previousListValueIds.contains(listValue.getId())));
+
+        List<ListValue<ConfidentialDivorceDocument>> sortedDocuments = new ArrayList<>();
+        sortedDocuments.addAll(0, documentsWithoutIds.get(true)); // add new documents to start of the list
+        sortedDocuments.addAll(1, documentsWithoutIds.get(false));
+
+        sortedDocuments.forEach(
+            uploadedDocumentListValue -> uploadedDocumentListValue.setId(String.valueOf(UUID.randomUUID()))
+        );
+
+        this.setConfidentialDocumentsUploaded(sortedDocuments);
     }
 
     private List<ListValue<DivorceDocument>> sortDocuments(final Map<Boolean, List<ListValue<DivorceDocument>>> documentsWithoutIds) {
